chore: Rename user-management to auth
