+++ auth/user-authentication-middleware.go
@@ -1,5 +1,5 @@
 // Author: Maximilian Floto
-package user_management
+package auth
 
 import (
 	"database/sql"
